Move isAttributeNameSafe to DOMProperty (#11802)
